# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Smart City Digital Citizen Portal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /garbage:
    get:
      operationId: listGarbageRequests
      tags: [garbage]
      summary: List garbage pickup requests
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [scheduled, in_progress, completed, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Garbage requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GarbageListResponse"
    post:
      operationId: createGarbageRequest
      tags: [garbage]
      summary: Request garbage pickup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GarbageInput"
      responses:
        "201":
          description: Garbage pickup request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GarbageRequest"
  /garbage/{id}:
    get:
      operationId: getGarbageRequest
      tags: [garbage]
      summary: Get garbage request by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Garbage request details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GarbageRequest"
    patch:
      operationId: updateGarbageRequest
      tags: [garbage]
      summary: Update garbage request status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GarbageUpdate"
      responses:
        "200":
          description: Updated garbage request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GarbageRequest"

  # ─────────────────────────────── PARKING ───────────────────────────────
  /parking/lots:
    get:
      operationId: listParkingLots
      tags: [parking]
      summary: List all parking lots
      parameters:
        - name: available
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Parking lots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ParkingLot"
  /parking/reservations:
    get:
      operationId: listParkingReservations
      tags: [parking]
      summary: List user parking reservations
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, cancelled]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Parking reservations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingReservationListResponse"
    post:
      operationId: createParkingReservation
      tags: [parking]
      summary: Make a parking reservation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ParkingReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingReservation"
  /parking/reservations/{id}:
    get:
      operationId: getParkingReservation
      tags: [parking]
      summary: Get parking reservation by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Parking reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingReservation"
    delete:
      operationId: cancelParkingReservation
      tags: [parking]
      summary: Cancel a parking reservation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reservation cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ParkingReservation"

  # ─────────────────────────────── TRANSPORT ───────────────────────────────
...
components:
  schemas:
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        totalPages:
          type: integer
      required: [total, page, limit, totalPages]
...
    GarbageRequest:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        address:
          type: string
        wasteType:
          type: string
          enum: [household, recyclable, hazardous, electronic, bulk]
        scheduledDate:
          type: string
          format: date
        timeSlot:
          type: string
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        notes:
          type: ["string", "null"]
        driverName:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, address, wasteType, scheduledDate, timeSlot, status, createdAt]
    GarbageInput:
      type: object
      properties:
        address:
          type: string
        wasteType:
          type: string
          enum: [household, recyclable, hazardous, electronic, bulk]
        scheduledDate:
          type: string
          format: date
        timeSlot:
          type: string
        notes:
          type: string
      required: [address, wasteType, scheduledDate, timeSlot]
    GarbageUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [scheduled, in_progress, completed, cancelled]
        driverName:
          type: string
    GarbageListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/GarbageRequest"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]

    # ─── Parking ───
    ParkingLot:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        address:
          type: string
        latitude:
          type: number
        longitude:
          type: number
        totalSpots:
          type: integer
        availableSpots:
          type: integer
        pricePerHour:
          type: number
        openingTime:
          type: string
        closingTime:
          type: string
        amenities:
          type: array
          items:
            type: string
        isActive:
          type: boolean
      required: [id, name, address, latitude, longitude, totalSpots, availableSpots, pricePerHour, openingTime, closingTime, amenities, isActive]
    ParkingReservation:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        parkingLotId:
          type: string
        parkingLot:
          $ref: "#/components/schemas/ParkingLot"
        vehicleNumber:
          type: string
        vehicleType:
          type: string
          enum: [car, motorcycle, truck, bicycle]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [active, completed, cancelled]
        totalAmount:
          type: number
        spotNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, userId, parkingLotId, parkingLot, vehicleNumber, vehicleType, startTime, endTime, status, totalAmount, createdAt]
    ParkingReservationInput:
      type: object
      properties:
        parkingLotId:
          type: string
        vehicleNumber:
          type: string
        vehicleType:
          type: string
          enum: [car, motorcycle, truck, bicycle]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
      required: [parkingLotId, vehicleNumber, vehicleType, startTime, endTime]
    ParkingReservationListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ParkingReservation"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"
      required: [data, pagination]

    # ─── Transport ───
...